+++ gp/vec.go
@@ -41,14 +41,14 @@ func (v Vec) Magnitude() float64 {
 }
 
 func (v Vec) X() float64 {
-	return C.gpVec_X(v.Vec)
+	return float64(C.gpVec_X(v.Vec))
 }
 
 func (v Vec) Y() float64 {
-	return C.gpVec_Y(v.Vec)
+	return float64(C.gpVec_Y(v.Vec))
 }
 
 func (v Vec) Z() float64 {
-	return C.gpVec_Z(v.Vec)
+	return float64(C.gpVec_Z(v.Vec))
 }
